vmwgfx/ttm: Add the set / drop master implementations.
This optimizes multi-master operations while allowing also for
multi-master operation and vt switching when operating on top of vesafb.

Assuming that any buffer validation takes place with the ttm_lock held
either in read- or write mode, this blocks clients trying to validate buffers
when their master is dropped. Clients belonging to other masters are not
affected. This is implemented using per-master ttm locks.

If the drm connection that was last master is dropped and then closed
without a new master is set, its clients waiting on the ttm lock will
receive a SIGTERM, making sure that we never deadlock dri clients waiting
for a new master to appear.

In this context, fbdev will act as a virtual drm master which is set when
no other master is set.

In addition, when we're operating in stealth mode on top of vesafb,
VRAM is cleared before the fbdev master is set,
making sure that we never interfere with vesafb.

For suspend and resume,
the current master ttm lock will be taken in suspend mode prior
to evicting buffers from volatile memory. This is not implemented yet.

The old vt switch ioctl is removed.

Signed-off-by: Thomas Hellstrom <thellstrom@vmware.com>