vmwgfx: Don't switch off the FIFO at master drop if there are 3D resources alive.
The FIFO is needed to destroy the 3D resources properly. Furthermore,
the SVGA device forgets all 3D state if the FIFO is turned off.

When Master is dropped, in addition TRACES is turned on in true stealth mode.
If vesafb is running, this means that the output will show correctly even if
the FIFO is turned on.

If Master is dropped while 3D resources are alive, and VGA is meant to take
over, that will, however, fail.

Signed-off-by: Thomas Hellstrom <thellstrom@vmware.com>