vmwgfx: Temporary fix for missing irqs
This is part of a larger, already pushed upstream fix that we can't
push to this repo yet, because we don't have drm event functionality.

Fixes a race condition caused by us reading the current fence seqno before
acknowledging the ANY_FENCE irq. The fence seqno could be updated after the
read but before acknowledging the irq, causing us to miss an irq.

Signed-off-by: Thomas Hellstrom <thellstrom@vmware.com>