vmwgfx: Allow legacy surfaces to have no id in destroy
Fixes erroneous BUG_ON on surface cleanup after resource
starvation error in execbuf ioctl. This happens because
resources structures are first allocated and then created
which populates the resource id, should userspace over-
commit on resources not all vmw_resources will be
populated and therefore destroy will be called on a non-
created surface.

Doing it this way seems to be okay, the GBS path handles
this case similarly.

Signed-off-by: Jakob Bornecrantz <jakob@vmware.com>
Reviewed-by: Biran Paul <brianp@vmware.com>
Reviewed-by: Zack Rusin <zackr@vmware.com>