vmwgfx: Clean up DMA mapping logic
Clean up the DMA mapping logic so that we use physical addresses by default
unless we detect IOMMU presence.
The user can override this in various ways:

force_dma_api = 1: Always map for DMA. Resort to coherent pages if needed.
If this isn't possible, disable DMA.
force_coherent = 1: Force use of coherent TTM pages. If this isn't possible,
disable DMA.

Also fix a thinko when mapping for DMA. We never need to do that before
the BIND operation regardless of mode, so remove the conditional populate
map for DMA call.

Signed-off-by: Thomas Hellstrom <thellstrom@vmware.com>
Reviewed-by: Brian Paul <brianp@vmware.com>
Reviewed-by: Jakob Bornecrantz <jakob@vmware.com>