vmwgfx: Use a customized iterator to get DMA addresses
Currently we store DMA addresses in a Scatter-Gather list and use the
scatter-gather list page iterator to get the addresses. However,
SG lists waste memory when we use physical addresses or coherent pages,
and also turns out to be tricky to set up when we use coherent pages.

Therefore, skip the SG list in favour of page arrays or dma address arrays
in these cases, and customize the page walking iterator to account for that.

Signed-off-by: Thomas Hellstrom <thellstrom@vmware.com>
Reviewed-by: Brian Paul <brianp@vmware.com>
Reviewed-by: Jakob Bornecrantz <jakob@vmware.com>