drm/vmwgfx: Add multimon screen target support
In this implementation, a Screen Target Display Unit tracks two
surfaces:  content and display.  Display is the buffer bound to
a screen target.  If the mode, i.e. screen target, is the same size
as the content buffer, then the display surface simply points back
to the content surface.

If the mode is not the same size as the content buffer, then a surface
is allocated specifically for display, and its content is blitted over
from the content buffer.

We can also share the same content buffer between two screen targets,
i.e. clone/mirrored mode.

* Move vmw_user_surface definition so STDU code can use it
* Removed vmw_screen_target_display because we don't need it

v2
Fixes to address review comments
* Removed code to pin/unpin content surface
* Coding standards relating to kernel comments
* Added support for DMA buffers
	* added vmw_kms_stdu_present()
	* modified set_config to handle DMA buffers
	* Instead of tracking content surface, track content fb
          because that's common between surfaces and DMA buffers
* Changed 8bpp handling to take screen targets into account.

Signed-off-by: Sinclair Yeh <syeh@vmware.com>
Reviewed-by: Thomas Hellstrom <thellstrom@vmware.com>