vmwgfx: Get rid of the vmwgfx sysfs subsystem
Instead reuse the drm subsystem created by the upstream drm module.
This commit creates a dependency on the upstream drm module.

The pm callbacks from the drm class are avoided, which makes sure that the
structure layout incompatibilities for the drm device and drm connector
doesn't matter.

This way we can avoid separate udev rules for the vmwgfx class, and also
user-space sysfs-aware drm clients can talk to a single class / subsystem.

Signed-off-by: Thomas Hellstrom <thellstrom@vmware.com>
Reviewed-by: Jakob Bornecrantz <jakob@vmware.com>